requirements.txt
streamlit
pandas
requests
langchain
langchain_google_genai
python-dotenv
